Update - Move DOM listeners from popups into eventMethods module
