+++ source/templates/chatStatusEntry.html
@@ -1,4 +1,4 @@
-<div class="flex-container flex-center stat-us-maximus-entry hover-highlight">
+<div class="flex-container flex-center stat-us-maximus-entry-row hover-highlight">
     <div class="flex-container flex-center-row gap-10px fs-90p text-muted w-100 px-10px">
         <div class="fa-solid kill-switch fs-lr cursor-pointer" title="Toggle entry's active state." data-i18n="[title]Toggle entry's active state."></div>
         <div class="fa-solid fa-eye-slash cursor-pointer private-lamp"></div>
