+++ source/templates/popupStatusEntry.html
@@ -1,4 +1,4 @@
-<div class="stat-us-maximus-popup-row flex-container flex-column flex-center-col gap-5px inline-drawer w-100 bordered-box">
+<div class="stat-us-maximus-popup-row stat-us-maximus-entry-row flex-container flex-column flex-center-col gap-5px inline-drawer w-100 bordered-box">
     <div class="inline-drawer-header flex-center gap-5px p-0 w-100">
         <span class="drag-handle fa-solid fa-bars"></span>
         <div class="inline-drawer-toggle fa-fw fa-solid fa-circle-chevron-down inline-drawer-icon down interactable"></div>
